fix exception on modal click_away
Previously a click away on the modal would trigger the close event. The close
event does a push_patch to return_to even when return_to is not defined.

According to the documentation for live_helpers, either return_to or
phx-target is required for Modal. However, prior to this commit phx-click
was set to id, and id was set to `:modal` in live_modal. And the modal's
handle_event for "close" expected a return_to.